Fix gcc warnings
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@4948 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ vision/star_detector/src/pydetector.cpp
@@ -38,7 +38,7 @@ PyObject *detect(PyObject *self, PyObject *args)
     std::vector<Keypoint> kp;
     sd->psd.DetectPoints(sd->img, std::back_inserter(kp));
     PyObject *r = PyList_New(kp.size());
-    for (int i = 0; i < kp.size(); i++) {
+    for (size_t i = 0; i < kp.size(); i++) {
         PyObject *t = PyTuple_New(4);
         PyTuple_SetItem(t, 0, PyInt_FromLong(kp[i].x));
         PyTuple_SetItem(t, 1, PyInt_FromLong(kp[i].y));
